refactor: move all folders to /app dir
