+++ app/components/profile-section/profile-section.tsx
@@ -11,9 +11,9 @@ import { useState } from 'react'
 import { toFormikValidationSchema } from 'zod-formik-adapter'
 
 import { DeleteUserDialog } from '@/app/components/profile-section/components/delete-user-dialog'
-import { UserUpdateInputSchema } from '@/schemas/user'
-import { UserUpdateInput } from '@/types/user'
-import { trpc } from '@/utils/trpc'
+import { UserUpdateInputSchema } from '@/app/schemas/user'
+import { UserUpdateInput } from '@/app/types/user'
+import { trpc } from '@/app/utils/trpc'
 
 export function ProfileSection() {
   const user = trpc.user.getLoggedIn.useQuery()
