+++ app/components/retro-view/components/retro-action-bar.tsx
@@ -6,7 +6,7 @@ import { useState } from 'react'
 import { ActionButtons } from '@/app/components/retro-view/components/action-bar/components/action-buttons/action-buttons'
 import { PhaseIndicator } from '@/app/components/retro-view/components/action-bar/components/phase-indicator/phase-indicator'
 import { RetroTimer } from '@/app/components/retro-view/components/action-bar/components/retro-timer/retro-timer'
-import { trpc } from '@/utils/trpc'
+import { trpc } from '@/app/utils/trpc'
 
 type RetroActionBarProps = {
   selectedRetro: Retrospective
