+++ app/components/retro-view/retro-view.tsx
@@ -7,7 +7,7 @@ import GridLoader from 'react-spinners/GridLoader'
 import { DiscussingView } from '@/app/components/retro-view/views/discussing-phase/discussing-view'
 import { VotingView } from '@/app/components/retro-view/views/voting-phase/voting-view'
 import { WritingView } from '@/app/components/retro-view/views/writing-phase/writing-view'
-import { trpc } from '@/utils/trpc'
+import { trpc } from '@/app/utils/trpc'
 
 export function RetroView() {
   const { resolvedTheme } = useTheme()
