+++ app/types/retro-item.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 
-import { RetroItemCreateInputSchema, RetroItemUpdateInputSchema } from '@/schemas/retro-item'
+import { RetroItemCreateInputSchema, RetroItemUpdateInputSchema } from '@/app/schemas/retro-item'
 
 export type RetroItemCreateInput = z.infer<typeof RetroItemCreateInputSchema>
